conversational-toolkit/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "conversational_toolkit"
authors = []
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.75.0,<0.116.0",
    "httpx>=0.11.0,<0.29.0",
    "python-jose>=1.3.0,<4.0.0",
    "loguru>=0.1.0,<1.0.0",
    "numpy>=1.0.0,<3.0.0",
    "docling>=2.0.0",
    "markitdown>=0.0.1",
    "openpyxl>=3.1.0",
    "rank-bm25>=0.2.2",
    "ragas>=0.2.0",
    "qwen-vl-utils==0.0.14",
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = { attr = "conversational_toolkit.__version__" }

[tool.ruff]
extend-include = ["*.ipynb"]
target-version = "py310"
line-length = 120
fix = true


[tool.ruff.lint]
extend-select = [
    "W605", # pycodestyle: invalid-escape-sequence
    "S102", # flake8-bandit: exec-builtin
    "INP", # flake8-no-pep420
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "PGH", # pygrep-hooks
    "PL", # Pylint
    "NPY", # NumPy-specific rules
    "RUF", # Ruff-specific rules
]
ignore = [
    "NPY002",   # numpy-legacy-random
    "PGH003",   # type ignore for mypy
    "PLR0913",  # too many arguments (common in class __init__ signatures)
    "PLC0415",  # import not at top-level — intentional lazy imports in async methods
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["INP001"]
"[!t][!e][!s][!t][!s]*/**" = ["PT"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
ignore_missing_imports = true
exclude = "^(venv|examples)/"
